Jump off the deep end with Moose
Mouse is buggy, Moose is slow, bare Perl OO is about as nice as
eating glass. Go with the slow option, since the whole point of
this exercise is to eventually switch from CGI to persistent
processes. (Moose even with the very minimal use so far would be
adding something like 0.3s latency to every request -- so this
code can not realistically be deployed as-is)