Add a 2nd password entry field to password reset page.
Not because I think that forcing the users to enter the password
again when changing it is a good idea. No, it's to prevent the
braindead browsers from autofilling this form with only data that should
not be there (a username instead of email address, a probably invalid
password).

Jump through the hoops, jump! Uncle Chrome knows what's best for you.